Feedback for loading large event logs
Having no feedback at all on progress, generally makes myself impatience so I suggest that you provide some feedback via a progress bar. The tqdm library is a useful quick and easy progress bar and offers lots of support for python environments from the command line to Jupiter notebooks, in which it can self identify which to use.

I do look for traces in the event log so this could add more time to loading but if I was working on a log with 100,000 traces, I think providing feedback would be worth the extra time to load the Eventlog object. The tqdm object is a decorator for any iterable but as generators don't have a __len__ function it provides only information on the speed of progress rather than the overall progress.

This concept could also be applied to the exporting as well.