Use input parameters instead of defaults
Issue: Parameters that are passed are not used sometimes and instead defaults are used. For non-default names for event_activity, event_timestamp and object_type this results in an error. Parameters are also not passed onto the OCEL object, which leads to defaults being used there as well.

Solution: Use parameter variables instead. Also parameters are passed onto the OCEL object, so now the instance variables can be assigned according to parameters.